finished debugging invalidResources test. found many problems with fake data
